+++ include/services/audio.h
@@ -14,6 +14,7 @@
 #include "SDL.h"
 #include "SDL_mixer.h"
 #include "resource_manager.h"
+#include "frame_timer.h"
 
 /**
  Singleton class for playing music and sound effects
@@ -23,7 +24,7 @@ private:
     Audio();
     ~Audio();
     Mix_Music *bg_track;
-    std::map<std::string, Uint32> last_played_map;
+    std::map<std::string, long> last_played_map;
 
     void setPlayed(const std::string &track_name);
 public:
@@ -32,7 +33,7 @@ public:
     void shutdown();
     void setBgTrack(const std::string &track_name);
     void playSound(const std::string &sound_name);
-    Uint32 getLastPlayed(const std::string &sound_name);
+    long getLastPlayed(const std::string &sound_name);
 };
 
 #endif /* audio_h */
